rename folder knative-function-controller -> funciton-controller
