+++ components/function-controller/Dockerfile
@@ -3,7 +3,7 @@ FROM golang:1.12.7 as builder
 #
 # Copy in the go src
 #
-WORKDIR /go/src/github.com/kyma-project/kyma/components/knative-function-controller
+WORKDIR /go/src/github.com/kyma-project/kyma/components/function-controller
 COPY pkg/    pkg/
 COPY cmd/    cmd/
 COPY vendor/ vendor/
@@ -11,7 +11,7 @@ COPY licenses/ licenses/
 #
 # Build
 #
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -a -o manager github.com/kyma-project/kyma/components/knative-function-controller/cmd/manager
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -a -o manager github.com/kyma-project/kyma/components/function-controller/cmd/manager
 #
 # Copy the controller-manager into a thin image
 #
@@ -21,8 +21,8 @@ WORKDIR /
 #
 # Copy binary
 #
-COPY --from=builder /go/src/github.com/kyma-project/kyma/components/knative-function-controller/manager .
-COPY --from=builder /go/src/github.com/kyma-project/kyma/components/knative-function-controller/licenses/ licenses/
+COPY --from=builder /go/src/github.com/kyma-project/kyma/components/function-controller/manager .
+COPY --from=builder /go/src/github.com/kyma-project/kyma/components/function-controller/licenses/ licenses/
 #
 # Run app
 #
