+++ components/function-controller/PROJECT
@@ -1,3 +1,3 @@
 version: "1"
 domain: kyma-project.io
-repo: github.com/kyma-project/kyma/components/knative-function-controller
+repo: github.com/kyma-project/kyma/components/function-controller
