+++ components/function-controller/cmd/manager/main.go
@@ -22,9 +22,9 @@ import (
 
 	buildv1alpha1 "github.com/knative/build/pkg/apis/build/v1alpha1"
 	servingv1alpha1 "github.com/knative/serving/pkg/apis/serving/v1alpha1"
-	"github.com/kyma-project/kyma/components/knative-function-controller/pkg/apis"
-	"github.com/kyma-project/kyma/components/knative-function-controller/pkg/controller"
-	"github.com/kyma-project/kyma/components/knative-function-controller/pkg/webhook"
+	"github.com/kyma-project/kyma/components/function-controller/pkg/apis"
+	"github.com/kyma-project/kyma/components/function-controller/pkg/controller"
+	"github.com/kyma-project/kyma/components/function-controller/pkg/webhook"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
