+++ components/function-controller/pkg/controller/function/function_controller_suite_test.go
@@ -26,7 +26,7 @@ import (
 
 	buildv1alpha1 "github.com/knative/build/pkg/apis/build/v1alpha1"
 	servingv1alpha1 "github.com/knative/serving/pkg/apis/serving/v1alpha1"
-	"github.com/kyma-project/kyma/components/knative-function-controller/pkg/apis"
+	"github.com/kyma-project/kyma/components/function-controller/pkg/apis"
 	"github.com/onsi/gomega"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
