+++ components/function-controller/pkg/utils/serving_utils_test.go
@@ -2,8 +2,8 @@ package utils_test
 
 import (
 	"github.com/ghodss/yaml"
-	serverlessv1alpha1 "github.com/kyma-project/kyma/components/knative-function-controller/pkg/apis/serverless/v1alpha1"
-	"github.com/kyma-project/kyma/components/knative-function-controller/pkg/utils"
+	serverlessv1alpha1 "github.com/kyma-project/kyma/components/function-controller/pkg/apis/serverless/v1alpha1"
+	"github.com/kyma-project/kyma/components/function-controller/pkg/utils"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"testing"
