+++ components/function-controller/pkg/webhook/default_server/server.go
@@ -46,7 +46,7 @@ func Add(mgr manager.Manager) error {
 		secretName = "webhook-server-secret"
 	}
 
-	svr, err := webhook.NewServer("knative-function-admission-server", mgr, webhook.ServerOptions{
+	svr, err := webhook.NewServer("function-admission-server", mgr, webhook.ServerOptions{
 		// TODO(user): change the configuration of ServerOptions based on your need.
 		Port:    9876,
 		CertDir: "/tmp/cert",
@@ -59,10 +59,10 @@ func Add(mgr manager.Manager) error {
 
 			Service: &webhook.Service{
 				Namespace: ns,
-				Name:      "knative-function-controller-webhook-server-service",
+				Name:      "function-controller-webhook-server-service",
 				// Selectors should select the pods that runs this webhook server.
 				Selectors: map[string]string{
-					"app":                     "knative-function-controller",
+					"app":                     "function-controller",
 					"control-plane":           "controller-manager",
 					"controller-tools.k8s.io": "1.0",
 				},
