+++ components/function-controller/pkg/webhook/default_server/webhook_suite_test.go
@@ -23,7 +23,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/kyma-project/kyma/components/knative-function-controller/pkg/apis"
+	"github.com/kyma-project/kyma/components/function-controller/pkg/apis"
 	"github.com/onsi/gomega"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
